Adding a new example NOT using "deadly_serious"
In order to show the overhead created by the serializing and
deserializing data between process. I made the same example, using
almost the same instructions, but using just one process and an array
instead of pipes.

In my box, that means something like: 50% of CPU use reduction AND
overall time for the same task.

SO, this implementation seems to have 2 main qualities: flexibility and
low memory use. "Performance" not being one of them.